feature: configuration file, beginning to refactor code
