+++ datarake/__main__.py
@@ -2,13 +2,15 @@ import argparse
 import csv
 import json
 import sys
+import yaml
 
-from datarake import DirectoryWalker, RakeMatch
-from datarake import RakeSet, RakeHostname, RakeEmail
-from datarake import RakeJWTAuth, RakeURL, RakePassword, RakeToken
-from datarake import RakeAWSHMACAuth, RakeBasicAuth, RakeBearerAuth
-from datarake import RakePrivateKey, RakeSSHIdentity, RakeNetrc, RakeSSHPass
-from datarake import RakePKIKeyFiles, RakeHtpasswdFiles, RakeKeystoreFiles
+from .common import DirectoryWalker
+from .common import RakeMatch
+from .common import RakeSet
+
+from .rakes import RakeContextPattern
+from .rakes import RakeFileMeta
+from .rakes import RakePattern
 
 class DataRakeWriter(object):
     '''
@@ -90,78 +92,6 @@ class DataRakeCSVWriter(DataRakeWriter):
         return
 
     def endOutput(self):
-        return
-
-class DataRakeInsightsWriter(DataRakeWriter):
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-        return
-
-    def initOutput(self):
-        print("{", end="", file=self._fd, flush=True)
-        self._count = 0
-        self._keys_written = 0
-        return
-
-    def initSecrets(self):
-        if self._quiet: return
-        print("\"vulnerabilities\": [", end="", file=self._fd, flush=True)
-        self._keys_written += 1
-        return
-
-    def writeSecret(self, secret):
-        if self._quiet: return
-        if self._count > 0:
-            print(",", end="", file=self._fd, flush=True)
-
-        line = secret.line if secret.line is not None else 0
-        startPos = secret.value_offset if secret.value_offset is not None else 0
-        length = secret.value_length if secret.value_length is not None else 0
-        endPos = startPos + length
-
-        vuln = { "endPos": endPos,
-                 "externalId": secret.external_id,
-                 "findingType": secret.label,
-                 "startLine": line,
-                 "endLine": line,
-                 "message": secret.description,
-                 "path": secret.file,
-                 "severity": secret.severity,
-                 "startPos": startPos }
-
-        jtxt = json.dumps(vuln)
-        print(jtxt, end="", file=self._fd, flush=True)
-        self._count += 1
-        return
-
-    def endSecrets(self):
-        if self._quiet: return
-        print("]", end="", file=self._fd, flush=True)
-
-        return
-
-    def initSummary(self):
-        if not self._summary: return
-
-        if self._keys_written > 0:
-            print(",", end="", file=self._fd, flush=True)
-
-        self._keys_written += 1
-        print("\"summary\": ", end="", file=self._fd, flush=True)
-        return
-
-    def writeSummary(self, s):
-        if not self._summary: return
-        print(json.dumps(s), end="", file=self._fd, flush=True)
-        return
-
-    def endSummary(self):
-        if not self._summary: return
-        # do nothing
-        return
-
-    def endOutput(self):
-        print("}", file=self._fd, flush=True)
         return
 
 
@@ -308,52 +238,12 @@ class DataRakeSARIFWriter(DataRakeWriter):
 def parseCmdLine(argv):
     parser = argparse.ArgumentParser()
 
-    # add OPTIONAL arguments
-    parser.add_argument("-n", "--hostname", action='store_true',
-                        required=False, default=False,
-                        help="scan for hostnames, optionally rooted in DOMAIN.")
-    parser.add_argument("-e", "--email", action='store_true',
-                        required=False, default=False,
-                        help="scan for email addresses, optionally rooted in DOMAIN.")
-    parser.add_argument("-d", "--domain", nargs=1,
-                        required=False, default=None,
-                        help="for hostname and emails, require that they are "
-                             "rooted in DOMAIN.  If no DOMAIN is specified and "
-                             "either hostname or email matching is enabled, any "
-                             "pattern matching a host or email will be reported")
-    parser.add_argument("-j", "--jwt", action='store_true',
-                        required=False, default=False,
-                        help="scan for Javascript Web Tokens (JWT)")
-
-    # allow DEFAULT arguments to be disabled.
-    parser.add_argument("-dp", "--disable-passwords", action='store_true',
-                        required=False, default=False,
-                        help="disable scan for passwords")
-    parser.add_argument("-dt", "--disable-tokens", action='store_true',
-                        required=False, default=False,
-                        help="disable scan for tokens")
-    parser.add_argument("-dh", "--disable-headers", action='store_true',
-                        required=False, default=False,
-                        help="disable scan for common auth headers")
-    parser.add_argument("-dk", "--disable-private-keys", action='store_true',
-                        required=False, default=False,
-                        help="disable scan for private key files.")
-    parser.add_argument("-du", "--disable-urls", action='store_true',
-                        required=False, default=False,
-                        help="disable scan for credentials in URLs")
-    parser.add_argument("-df", "--disable-dangerous-files", action='store_true',
-                        required=False, default=False,
-                        help="disable detection of dangerous files")
-    parser.add_argument("-dc", "--disable-dangerous-commands", action='store_true',
-                        required=False, default=False,
-                        help="disable detection of dangerous commands")
-
     parser.add_argument("PATH", default=["."], nargs="*",
                         help="Path to be (recursively) searched.")
 
     # output formatting
     parser.add_argument("-f", "--format", nargs=1, required=False, type=str,
-                        choices=["csv", "json", "insights", "sarif"], default=["csv"],
+                        choices=["csv", "json", "sarif"], default=["json"],
                         help="Output format")
     parser.add_argument("-o", "--output", nargs=1, required=False, type=str, default=None,
                         help="Output location (defaults to stdout)")
@@ -362,7 +252,7 @@ def parseCmdLine(argv):
     parser.add_argument("-dx", "--disable-context", required=False, action="store_true", default=False,
                         help="Disable output of context match")
     parser.add_argument("-dv", "--disable-value", required=False, action="store_true", default=False,
-                        help="Disable output of secret match")
+                        help="Disable output of secret matched")
     parser.add_argument("-u", "--summary", required=False, action="store_true", default=False,
                         help="enable output of summary statistics")
     parser.add_argument("-q", "--quiet", required=False, action="store_true", default=False,
@@ -373,63 +263,44 @@ def parseCmdLine(argv):
 
     return parser.parse_args(argv[1:])
 
+def loadConfig(cfile:str="etc/datarake.yaml"):
+    with open(cfile, "r") as fd:
+        cfg = yaml.safe_load()
+
+    rs = RakeSet(verbose=cfg.verbose)
+
+    for r in cfg['Rakes']:
+        if   r['type'] == "ContextPattern": c = RakeContextPattern()
+        elif r['type'] == "FileMeta":       c = RakeFileMeta()
+        elif r['type'] == "SimplePattern":  c = RakePattern()
+        else:
+            raise RuntimeError(f"ERROR: unsupported Rake type: {r['type']}")
+
+        rake = c()
+        c.load(r)
+        rs.add(rake)
+
+    return rs
+
 
 def main(argv=sys.argv):
     cfg = parseCmdLine(argv)
+    if cfg.secure: RakeMatch.set_secure()
+    if cfg.disable_context: RakeMatch.disable_context()
+    if cfg.disable_value: RakeMatch.disable_value()
+
+    rs = loadConfig(cfg.config_file)
+
     if cfg.output is None:
         fd = sys.stdout
     else:
         fd = open(cfg.output[0], 'w', encoding='utf-8')
 
-    if cfg.secure: RakeMatch.set_secure()
-    if cfg.disable_context: RakeMatch.disable_context()
-    if cfg.disable_value: RakeMatch.disable_value()
-
-    rs = RakeSet(verbose=cfg.verbose)
-
-    if cfg.hostname:
-        rs.add(RakeHostname(domain = cfg.domain))
-
-    if cfg.email:
-        rs.add(RakeEmail(domain = cfg.domain))
-
-    if cfg.jwt:
-        rs.add(RakeJWTAuth())
-
-    if not cfg.disable_urls:
-        rs.add(RakeURL())
-
-    if not cfg.disable_passwords:
-        rs.add(RakePassword())
-
-    if not cfg.disable_passwords:
-        rs.add(RakeToken())
-
-    if not cfg.disable_headers:
-        rs.add(RakeAWSHMACAuth())
-        rs.add(RakeBasicAuth())
-        rs.add(RakeBearerAuth())
-
-    if not cfg.disable_private_keys:
-        rs.add(RakePrivateKey())
-
-    if not cfg.disable_dangerous_files:
-        rs.add(RakeSSHIdentity())
-        rs.add(RakeNetrc())
-        rs.add(RakePKIKeyFiles())
-        rs.add(RakeHtpasswdFiles())
-        rs.add(RakeKeystoreFiles())
-
-    if not cfg.disable_dangerous_commands:
-        rs.add(RakeSSHPass())
-
     out_format = cfg.format[0]
 
-    # secure quiet summary
+    # verbosity:  secure quiet summary
     if out_format == 'csv':
         writer = DataRakeCSVWriter(fd=fd, quiet=cfg.quiet, summary=cfg.summary)
-    elif out_format == 'insights':
-        writer = DataRakeInsightsWriter(fd=fd, quiet=cfg.quiet, summary=cfg.summary)
     elif out_format == 'sarif':
         writer = DataRakeSARIFWriter(fd=fd, quiet=cfg.quiet, summary=cfg.summary)
     else:
